use createOrUpdate when creating marketo leads
Many duplicate leads have been created. Logs show that we might be
calling marketo up to 5x (or more?) for a newly created user. By
switching from createOnly to createOrUpdate, I hope to allow marketo
to deduplicate on the email address, which will hopefully minimize the
number of duplicates created.

The API docs are not clear about the circumstances under which
duplicates are created. They indicate only that when multiple requests
with the same "key" arrive in rapid succession, duplicates can
happen. They don't define the "key" however. I think it's logical to
assume that it's the "LookupField" parameter, which is ignored when
using createOnly, but should be valid when using createOrUpdate, hence
the change.

BACK-2999